--- a/cmd/kubenab/admissions.go
+++ b/cmd/kubenab/admissions.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"encoding/json"
 	"fmt"
 	"github.com/jfrog/kubenab/log"
 	"io/ioutil"
@@ -10,6 +9,7 @@ import (
 	"strconv"
 	"strings"
 
+	json "github.com/json-iterator/go"
 	"k8s.io/api/admission/v1beta1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"