--- a/cmd/kubenab/go.mod
+++ b/cmd/kubenab/go.mod
@@ -2,6 +2,7 @@ module github.com/jfrog/kubenab
 
 require (
 	github.com/gogo/protobuf v1.3.0 // indirect
+	github.com/json-iterator/go v1.1.7
 	github.com/prometheus/client_golang v1.1.0
 	github.com/prometheus/client_model v0.0.0-20190812154241-14fe0d1b01d4 // indirect
 	github.com/prometheus/common v0.7.0 // indirect