--- a/cmd/kubenab/log/log_debug.go
+++ b/cmd/kubenab/log/log_debug.go
@@ -1,0 +1,17 @@
+// +build !strip_debug
+
+package log
+
+import "log"
+
+func Debug(v ...interface{}) {
+	log.Print(v...)
+}
+
+func Debugln(v ...interface{}) {
+	log.Println(v...)
+}
+
+func Debugf(fmt string, args ...interface{}) {
+	log.Printf(fmt, args...)
+}