--- a/cmd/kubenab/log/log_release.go
+++ b/cmd/kubenab/log/log_release.go
@@ -1,0 +1,9 @@
+// +build strip_debug,!debug
+
+package log
+
+func Debug(v ...interface{}) {}
+
+func Debugln(v ...interface{}) {}
+
+func Debugf(fmt string, args ...interface{}) {}