### Currently Improvements

The average Latency could be decreased by 165,49855 % (from 759,14us to
285,93us).
The average Request/s improved by 62,057821 % (from 2620,34 req/s to 6906,14
req/s).

#### Benchmark Raw Data


Added the improvements in percentage (-n % is good in _Latency_!)

##### With `strip_debug`

```
Statistics        Avg      Stdev        Max
  Reqs/sec      6906,14     612,30    7957,42
  Latency      285,93us   171,41us    27,91ms
  Latency Distribution
     50%   257,00us   (-58,593750 %)
     75%   309,00us   (-60,517799 %)
     90%   376,00us   (-66,755319 %)
     95%   440,00us   (-70,909091 %)
     99%   764,00us   (-2839,9464 %)
  HTTP codes:
    1xx - 0, 2xx - 100000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:    15.05MB/s
```

##### Without `strip_debug`

```
Statistics        Avg      Stdev        Max
  Reqs/sec      2620,34    1478,93    5612,99
  Latency      759,14us     2,42ms    33,50ms
  Latency Distribution
     50%   406,00us
     75%   496,00us
     90%   627,00us
     95%   752,00us
     99%    21,95ms
  HTTP codes:
    1xx - 0, 2xx - 100000, 3xx - 0, 4xx - 0, 5xx - 0
    others - 0
  Throughput:     5.57MB/s
```