@l0nax 2019-10-14
The average Latency is improved – by the latest Patches – by _55,49855 %_ and the req/s by _10,057821 %_